requirements.txt
PyNaCl
requests
python-dotenv
internetarchive
tqdm
PyYAML
reportlab
